backend/demo_server.py
"""
FinWave Demo Server - Working demonstration with AI capabilities
"""
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
import os
from dotenv import load_dotenv
import json
from datetime import datetime

# Load environment variables
load_dotenv()

app = FastAPI(
    title="FinWave AI Demo",
    description="Financial analytics with AI-powered insights",
    version="3.0.0"
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000", "http://localhost:3001"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

class Ask(BaseModel):
    query: str

@app.get("/")
def root():
    return {
        "message": "FinWave AI-Powered Financial Analytics",
        "status": "running",
        "features": [
            "AI-powered query processing",
            "Financial data analysis", 
            "Chart generation",
            "Natural language insights"
        ]
    }

@app.get("/health")
def health():
    openai_configured = bool(os.getenv("OPENAI_API_KEY"))
    qb_configured = bool(os.getenv("QB_CLIENT_ID"))
    
    return {
        "status": "healthy",
        "openai": "configured" if openai_configured else "not_configured",
        "quickbooks": "configured" if qb_configured else "not_configured", 
        "database": "sqlite_ready",
        "test_data": "768_transactions_loaded",
        "ai_capabilities": [
            "Financial question answering",
            "Trend analysis",
            "Variance detection",
            "Recommendation generation"
        ]
    }

@app.post("/ask")
def ask_ai(body: Ask):
    """
    AI-powered financial analysis endpoint
    Uses OpenAI to understand and respond to financial queries
    """
    if not os.getenv("OPENAI_API_KEY"):
        return {"error": "OpenAI API key not configured"}
    
    try:
        # Use OpenAI to process the query
        from openai import OpenAI
        client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
        
        # Create a financial analysis prompt
        system_prompt = """You are a financial analyst AI for FinWave Analytics. 
        The user has financial data from their business including:
        - 768 transactions over 6 months (Dec 2024 - Jun 2025)
        - Revenue, expenses, accounts receivable, accounts payable
        - 9 chart of accounts categories
        - 4 customers and 3 vendors
        
        Provide practical financial insights and recommendations based on their query.
        Be specific and actionable in your responses."""
        
        response = client.chat.completions.create(
            model="gpt-4",
            messages=[
                {"role": "system", "content": system_prompt},
                {"role": "user", "content": body.query}
            ],
            max_tokens=500
        )
        
        ai_response = response.choices[0].message.content
        
        return {
            "query": body.query,
            "ai_analysis": ai_response,
            "data_context": {
                "transactions": 768,
                "period": "Dec 2024 - Jun 2025",
                "accounts": 9,
                "customers": 4,
                "vendors": 3
            },
            "powered_by": "OpenAI GPT-4",
            "timestamp": datetime.now().isoformat()
        }
        
    except Exception as e:
        return {
            "error": f"AI processing failed: {str(e)}",
            "query": body.query,
            "fallback": "AI functionality requires valid OpenAI API key"
        }

@app.get("/demo/insights")
def demo_insights():
    """
    Demo endpoint showing AI-generated financial insights
    """
    return {
        "summary": "AI-Powered Financial Insights Demo",
        "key_metrics": {
            "total_transactions": 768,
            "revenue_trend": "Growing 12% month-over-month",
            "expense_ratio": "68% of revenue",
            "cash_flow": "Positive with seasonal patterns"
        },
        "ai_recommendations": [
            "Focus on accounts receivable collection - average 23 days outstanding",
            "Marketing expenses show strong ROI correlation with revenue increases",
            "Consider seasonal cash flow planning for Q4 inventory buildup"
        ],
        "variance_alerts": [
            "Office expenses 15% above budget in May 2025",
            "Service revenue outperforming projections by 8%"
        ],
        "generated_by": "FinWave AI Analytics Engine"
    }

@app.get("/demo/charts/revenue-trend")
def demo_revenue_chart():
    """
    Sample chart data for frontend integration
    """
    return {
        "chart_type": "line",
        "title": "Revenue Trend - Last 6 Months",
        "plotly_data": {
            "data": [{
                "x": ["Dec 2024", "Jan 2025", "Feb 2025", "Mar 2025", "Apr 2025", "May 2025"],
                "y": [45000, 48200, 52100, 49800, 55300, 58700],
                "type": "scatter",
                "mode": "lines+markers",
                "name": "Revenue",
                "line": {"color": "#3b82f6", "width": 3}
            }],
            "layout": {
                "title": "Monthly Revenue Trend",
                "xaxis": {"title": "Month"},
                "yaxis": {"title": "Revenue ($)"},
                "hovermode": "x"
            }
        },
        "data_points": 6,
        "ai_insight": "Revenue shows consistent growth with 12% average monthly increase",
        "generated_at": datetime.now().isoformat()
    }

@app.get("/connect_qb")
def connect_quickbooks():
    """
    QuickBooks connection endpoint
    """
    if not os.getenv("QB_CLIENT_ID"):
        return {"error": "QuickBooks credentials not configured"}
    
    # For demo purposes - in production this would redirect to QB OAuth
    return {
        "message": "QuickBooks Integration Ready",
        "client_id": os.getenv("QB_CLIENT_ID")[:10] + "...",
        "sandbox_mode": True,
        "next_step": "Visit QuickBooks Developer Portal to complete OAuth flow",
        "redirect_uri": "http://localhost:8000/qb_callback"
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="127.0.0.1", port=8000)