backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
narwhals==1.42.1
numpy==2.3.0
openai==1.86.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.0
pgvector==0.4.1
plotly==6.1.2
psycopg==3.2.9
psycopg-binary==3.2.9
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
xxhash==3.5.0
zstandard==0.23.0
openai

# Excel processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# OAuth and authentication
oauthlib==3.2.2
requests-oauthlib==1.3.1

# CRM integrations
simple-salesforce==1.12.5
hubspot-api-client==8.2.1

# Authentication
pyjwt[crypto]==2.8.0
python-jose[cryptography]==3.3.0
auth0-python==4.7.1

# Database
alembic==1.13.1
duckdb==0.10.0
duckdb-engine==0.11.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0

# Monitoring (optional)
prometheus-client==0.19.0
